Make sems container a component
